create separate eventStream files
